+++ plugins/time/src/main/java/eu/automateeverything/timeplugin/NowValueNode.kt
@@ -7,6 +7,6 @@ import java.util.*
 class NowValueNode : ValueNode {
     override fun getValue(now: Calendar): PortValue {
         val nowSeconds = now.get(Calendar.HOUR_OF_DAY)*3600 + now.get(Calendar.MINUTE)*60 + now.get(Calendar.SECOND)
-        return SecondOfDayStamp(nowSeconds.toDouble())
+        return SecondOfDayStamp(nowSeconds.toBigDecimal())
     }
 }
